requirements.txt
python-chess==1.999
rich==15.0.0
pillow==12.2.0
numpy==2.4.4
edge-tts==7.2.8
